# kubernetes/test: fix slow tests