<!--

Please describe the motivation for this PR and provide enough
information so that others can review it.

-->
**Description**:
The tests had synchronization issues resulting from missing and blocking
event notifications.

The missing events were a result of the tests
not waiting till the namespace used in the test is monitored, which
in turn causes the caches for the other Kubernetes resources to
be filtered out as unmonitored. The result is that the resource is
only marked as monitored upon the next resync from k8s informers
(currently every 30s). This is fixed by ensuring an async wait
on the namespace to be monitored before waiting on k8s events for
resources belonging to the namespace.

Since unbuffered channels are used for announcements, it is better
to read as soon as data is available rather than to batch reads,
which results in writes being blocked. This is also addressed in
the tests.

This change speeds up by the test by over 60s to 90s when the events
are missed which was happening often.

It also removes some unnecessary code in the tests.

Resolves #2014

Before this change:
```
ok  	github.com/openservicemesh/osm/pkg/kubernetes	90.779s
ok  	github.com/openservicemesh/osm/pkg/kubernetes	90.769s
ok  	github.com/openservicemesh/osm/pkg/kubernetes	90.825s
ok  	github.com/openservicemesh/osm/pkg/kubernetes	90.787s
ok  	github.com/openservicemesh/osm/pkg/kubernetes	90.791s
ok  	github.com/openservicemesh/osm/pkg/kubernetes	90.794s
ok  	github.com/openservicemesh/osm/pkg/kubernetes	60.869s
ok  	github.com/openservicemesh/osm/pkg/kubernetes	90.797s
ok  	github.com/openservicemesh/osm/pkg/kubernetes	90.769s
ok  	github.com/openservicemesh/osm/pkg/kubernetes	90.796s
```

After this change:
```
ok  	github.com/openservicemesh/osm/pkg/kubernetes	1.123s
ok  	github.com/openservicemesh/osm/pkg/kubernetes	1.103s
ok  	github.com/openservicemesh/osm/pkg/kubernetes	1.099s
ok  	github.com/openservicemesh/osm/pkg/kubernetes	1.092s
ok  	github.com/openservicemesh/osm/pkg/kubernetes	1.102s
ok  	github.com/openservicemesh/osm/pkg/kubernetes	1.117s
ok  	github.com/openservicemesh/osm/pkg/kubernetes	1.099s
ok  	github.com/openservicemesh/osm/pkg/kubernetes	1.140s
ok  	github.com/openservicemesh/osm/pkg/kubernetes	1.096s
ok  	github.com/openservicemesh/osm/pkg/kubernetes	1.094s
```

<!--

Please mark with X for applicable areas.

-->
**Affected area**:

- New Functionality      [ ]
- Documentation          [ ]
- Install                [ ]
- Control Plane          [ ]
- CLI Tool               [ ]
- Certificate Management [ ]
- Networking             [ ]
- Metrics                [ ]
- SMI Policy             [ ]
- Security               [ ]
- Tests                  [X]
- CI System              [ ]
- Performance            [ ]
- Other                  [ ]


Please answer the following questions with yes/no.

- Does this change contain code from or inspired by another project? If so, did you notify the maintainers and provide attribution?
`No`